Disable major version check for publishing process
Commented out the conditional check in the `PushPackages` class that verifies if the major version of `GitVersionInformation` is >= 1. This change disables the execution of `context.ApaxPublish()` based on the major version, indicating a shift in the publishing process logic.

Update artifact management and publishing logic

Modified `CreateArtifactsTask` to include an additional path `"assets"` in the `ApaxChangeBuildProperties` method. Commented out the major version check in `PushPackages`, which prevents the `ApaxPublish` method from being called based on that condition. These changes streamline artifact handling and adapt to new directory structures.