Refactor details rendering logic
Refactored the details rendering logic by removing the `detailsRcc` property and its associated initialization and logic. Replaced it with `DetailsContext` and `DetailsPresentationType` properties to streamline state management. Updated `RenderableContentControlFragment` to use the new properties.

Introduced a new `OpenDetails` method to handle setting the context and presentation type, and toggling the modal window. Removed unused and commented-out code for better maintainability.